% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movies}
\alias{movies}
\alias{movies_test}
\alias{movies_train}
\title{FiveThirtyEight's Movie Scores}
\format{A \code{data.frame} with 146 rows and 9 columns. The training data
  contains 87 movies, while the test set contains 59 movies.}
\source{
\url{https://github.com/fivethirtyeight/data/tree/master/fandango}
}
\usage{
movies

movies_train

movies_test
}
\description{
This dataset contains every film that has a Rotten Tomatoes
  rating, a RT User rating, a Metacritic score, a Metacritic User score,
  and IMDb score, and at least 30 fan reviews on Fandango. The data from
  Fandango was pulled on Aug. 24, 2015.
}
\keyword{datasets}

