% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/summary.PUGS.R
\name{summary.PUGS}
\alias{summary.PUGS}
\title{Gather samples of predictions}
\usage{
\method{summary}{PUGS}(object, ...)
}
\arguments{
\item{object}{A \code{pugs} object generated by \code{\link{predict.ECC}}.}

\item{...}{\code{type = "prob"} for probabilistic predictions,
\code{type = "class"} for binary (0/1) predictions}
}
\value{
A matrix of predictions.
}
\description{
Gather samples of predictions
}
\examples{
\dontrun{
  y.pred <- train_ecc(x, y) \%>\% predict(x.new) \%>\% summary
}
}

