% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/train_ecc.R
\name{train_ecc}
\alias{train_ecc}
\title{Train an Ensemble of Classifier Chains (ECC)}
\usage{
train_ecc(x, y, m = 1, parallel = TRUE, silent = FALSE, FUN = NULL, ...)
}
\arguments{
\item{x}{A data frame or matrix of features.}

\item{y}{A data frame or matrix of labels. Each label must be its own column
and each instance (observation) must be a row of 0s and 1s, indicating which
labels belong to the instance.}

\item{m}{Number of classifier chains (models) to train. Recommended:
\code{m = 3} and \code{m = 7} for 4-core and 8-core systems, respectively.}

\item{parallel}{Whether to utilize multicore capabilities of the system.}

\item{silent}{Whether to print progress messages to console. Recommended.}

\item{FUN}{User-supplied classifier training function. If not supplied, the
trainer will use the built-in classifier. See Details for more information.}

\item{...}{additional arguments to pass to \code{FUN}.}
}
\value{
A object of class \code{ecc}.
}
\description{
Train an Ensemble of Classifier Chains (ECC)
}
\examples{
\dontrun{
  fit <- train_ecc(x, y, FUN = randomForest::randomForest)
  fit <- train_ecc(x, y, m = 7, FUN = C50::C5.0, trials = 10)
}
}

