% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Assess prediction accuracy}
\usage{
validate(object, y)
}
\arguments{
\item{object}{A \code{PUGS} object generated by \code{\link{predict.ECC}}.}

\item{y}{A matrix of the same form as the one used with
\code{\link{train_ecc}}.}
}
\value{
A variety of multi-label classification accuracy measurements.
}
\description{
Assess prediction accuracy
}
\examples{
\dontrun{
  train_ecc(x.train, y.train) \%>\% predict(x.test) \%>\% validate(y.test)
}
}

