% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugs.R
\name{summary.PUGS}
\alias{summary.PUGS}
\title{Gather samples of predictions}
\usage{
\method{summary}{PUGS}(object, ...)
}
\arguments{
\item{object}{A \code{pugs} object generated by \code{\link{predict.ECC}}.}

\item{...}{\code{type = "prob"} for probabilistic predictions,
\code{type = "class"} for binary (0/1) predictions}
}
\value{
A matrix of predictions.
}
\description{
Collapses the multi-label predictions across sets of classifier
  chains and iterations into a single set of predictions, either binary or
  probabilistic.
}

