% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pugs.R
\name{validate_pugs}
\alias{validate_pugs}
\title{Assess multi-label prediction accuracy}
\usage{
validate_pugs(object, y)
}
\arguments{
\item{object}{A \code{PUGS} object generated by \code{\link{predict.ECC}}.}

\item{y}{A matrix of the same form as the one used with
\code{\link{ecc}}.}
}
\value{
A variety of multi-label classification accuracy measurements.
}
\description{
Computes a variety of accuracy metrics for multi-label
  predictions.
}
\examples{
\dontrun{
x <- movies_train[, -(1:3)]
y <- movies_train[, 1:3]

model_c50 <- ecc(x, y, .f = C50::C5.0)
predictions_c50 <- predict(model_c50, movies_test[, -(1:3)],
                           .f = C50::predict.C5.0, type = "prob")
validate(predictions_c50, movies_test[, 1:3])
  
model_rf <- ecc(x, y, .f = randomForest::randomForest)
predictions_rf <- predict(model_rf, movies_test[, -(1:3)],
                          n.iters = 1000, burn.in = 100, thin = 10,
                          .f = function(rF, newdata){
                            randomForest:::predict.randomForest(rF, newdata, type = "prob")
                          })
validate(predictions_rf, movies_test[, 1:3])
}
}

